back/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.11
fastembed==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-openai==0.3.8
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.6
langsmith==0.3.13
loguru==0.7.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.1
onnxruntime==1.21.0
openai==1.66.3
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
py_rust_stemmers==0.1.5
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==309
PyYAML==6.0.2
qdrant-client==1.13.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.15.2
seaborn==0.13.2
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.90
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
win32_setctime==1.2.0
yarl==1.18.3
zstandard==0.23.0
